fixes failing tests
