+++ test/back/endpoints/endpoints.test.js
@@ -187,11 +187,7 @@ test('/authenticate-user endpoint returns error for delAsync Redis call', (t) =>
     {
         method: 'post',
         url: '/save-user',
-        payload: {
-            email: 'lecturer@city.ac.uk',
-            password: 'testinglecturer',
-            is_lecturer: true
-        },
+        payload: franzCreds,
         expected: {
             message: 'user exists'
         }
@@ -222,36 +218,6 @@ test('/authenticate-user endpoint returns error for delAsync Redis call', (t) =>
         },
         expected: {
             message: 'The code you have entered is invalid'
-        }
-    },
-    {
-        method: 'post',
-        url: '/save-user',
-        payload: {
-            email: 'sohilpandya1990@gmail.com',
-            password: 'testingstudent',
-            is_lecturer: false,
-            username: 'testingstudent',
-            group_code: ''
-        },
-        expected: {
-            email: 'sohilpandya1990@gmail.com',
-            expiry_code: null,
-            is_lecturer: false,
-            is_verified: true,
-            reset_password_code: null,
-            user_id: 44,
-            username: 'testingstudent',
-            group_code: null,
-            verification_code: null,
-            is_super_admin: false,
-            trial_expiry_time: null,
-            account_type: null,
-            paid: null,
-            group_admin_has_paid: null,
-            is_group_admin: false,
-            is_user_active: true,
-            moodle_id: null
         }
     },
     {
